+++ index.html
@@ -63,6 +63,6 @@
     <p>&copy;Copyright 2023 <a href=""> The Readers </a> - All Rights Reserved</p>
   </footer>
 
-  <script src="app.js"></script>
+  <script type="module" src="/index.js"></script>
 </body>
 </html>
